refactor(tests): consolidate monotonically_increasing_id tests into a single file
